requirements.txt
streamlit
pillow
pdf2image
boto3
phonenumbers
awscli
json
botocore.exceptions
requests